fixed multiple deletions in payment registers. preparations for clearing code and using utils fetches along with es7 functions
